Move geo-projection component into a separate file
